Move schema adapter tests
